refactor: Move type hierarchy to analysis module and gate the whole ir module
