+++ src/analysis/type_hierarchy/petgraph_impl.rs
@@ -8,7 +8,7 @@ use petgraph::{
 };
 
 use crate::{
-    ir::{ClassHierarchy, InterfaceImplHierarchy},
+    analysis::{ClassHierarchy, InterfaceImplHierarchy},
     jvm::references::ClassRef,
 };
