+++ src/ir/control_flow/mod.rs
@@ -53,7 +53,7 @@ impl<N, E> ControlFlowGraph<N, E> {
         ProgramCounter::ZERO
     }
 
-    /// Transforms the node and edge data to construt a new control flow graph.
+    /// Transforms the node and edge data to construct a new control flow graph.
     #[must_use]
     pub fn map<N1, E1, NMap, EMap>(self, nf: NMap, ef: EMap) -> ControlFlowGraph<N1, E1>
     where
