+++ src/ir/petgraph/mod.rs
@@ -1,4 +1,3 @@
 //! Implementations for graphs.
 pub mod cfg;
 pub mod du_chain;
-pub mod type_hierarchy;
