+++ src/jvm/bytecode/mod.rs
@@ -17,10 +17,10 @@ use std::{
     num::TryFromIntError,
 };
 
+#[cfg(not(feature = "unstable-bytecode-generation"))]
+pub(crate) use errors::GenerationError;
 #[cfg(feature = "unstable-bytecode-generation")]
 pub use errors::{GenerationError, GenerationErrorKind};
-#[cfg(not(feature = "unstable-bytecode-generation"))]
-pub(crate) use errors::{GenerationError, GenerationErrorKind};
 pub use errors::{ParseError, ParseErrorKind};
 use num_traits::ToBytes;
